+++ Car/include/Car/internal/Vulkan/SSBO.hpp
@@ -5,16 +5,16 @@
 namespace Car {
     class VulkanSSBO : public SSBO {
     public:
-        VulkanSSBO(void* data, uint32_t size, Buffer::Usage usage);
+        VulkanSSBO(void* data, uint64_t size, Buffer::Usage usage);
         virtual ~VulkanSSBO() override;
 
-        virtual void updateBuffer(void* data, uint32_t size) override;
+        virtual void updateBuffer(void* data, uint64_t size) override;
 
-        virtual uint32_t getSize() const override { return mSize; }
+        virtual uint64_t getSize() const override { return mSize; }
         virtual void bind(uint32_t binding) const override;
         virtual void unbind() const override;
 
     private:
-        uint32_t mSize;
+        uint64_t mSize;
     };
 } // namespace Car
